Rewrite Disclosure and DisclosureGroup
