+++ packages/bento-design-system/src/index.ts
@@ -19,8 +19,8 @@ export * from "./Chip/Chip";
 export * from "./ContentBlock/ContentBlock";
 export * from "./ContentWithSidebar/ContentWithSidebar";
 export * from "./DateField/DateField";
-export * from "./Disclosure/createDisclosure";
-export * from "./DisclosureGroup/createDisclosureGroup";
+export * from "./Disclosure/Disclosure";
+export * from "./DisclosureGroup/DisclosureGroup";
 export * from "./Divider/Divider";
 export * from "./Feedback/createFeedback";
 export * from "./Field/Field";
